refactor: 사용자 redux 상태명 변경
